requirements.prod.txt
fastapi
uvicorn
pydantic
xgboost==1.7.6
scikit-learn
pandas
numpy
shap
imbalanced-learn
streamlit
plotly
requests
evidently==0.4.16
optuna
python-multipart
sqlalchemy
psycopg2-binary